return raw value for ControlType and RuntimeId
For ControlType and RuntimeId in ScriptProperty, the value obtained from
AutomationElement was converted to a string (for display purpuse), but
in order to maintain consistency with other property values, change to
return the value as it is.
The RuntimeId used for the provider path requires a string format that
does not contain spaces, so change the name to ItemId and make it a
different ScriptProperty.